frontend: Add toast messages on successful saving and navigation
Before the user was simply redirected to the details page without any indication that the save was successful.
Now a toast message is shown for the user after the user is redirected to the details page.
This could potentially be used to display error messages also, instead of a modal which requires a click to close