frontend: Fix an error on the first time creating an apartment
When creating an apartment resulted in an error due to street address not being included in the request,
street address was really added to the state, but the state didn't have enough time to be updated which caused the error.

Use the local formData instead of state formData when sending a request to create an apartment.