Add ability to add conditions of sale for additional ownerships
This can be used to create conditions of sale for ownerships that the
owner does not own theirself, but through some other relationship outside
of the system, like sharing a household.

Conditions of sale are only created to these ownerships if they are
for new apartments, but not between another additional ownership if
multiple are defined at the same time.

This naturally removes the requirement for condition of sale to have
the same owner in their ownerships.