Improve apartment.is_new deduction
Apartments are now considered new after the first sale until they no
longer have any conditions of sale form their first sale. This allows
additional conditions of sale to be created for them based on households,
even if the apartment has been completed before the first sale.

There is a small edge case that is not covered: two new apartments only
have one condition of sale between them, and this does not count in
deducting apartment newness for one of the apartments that is new. So,
if that is the only condition of sale for that apartment, it wouldn't
be considered new, however, it is very unlikely that this would cause
any issues, as the owner likely has old apartments in this case as well.