Change sale logic so that previous ownerships are deleted on sale
This allows conditions of sale to be automatically fulfilled when
a new sale is made, and makes it easier to find owners which no longer
own any regulated histas apartments for obfuscation.

The semantics of the Owner queryset are also nicer, as
Ownership.objects.filter(owner=owner) now returns all current ownerhips
for that owner, and not all ownerships it has ever had (this can still
be retrieved with Ownership.all_objects.filter(owner=owner)).